fix date diaplay issue
