+++ src/components/fields/edit/DateEditor.jsx
@@ -26,21 +26,7 @@ export default function DateEditor(props) {
   const intl = useIntl();
   let dateValue =   
     typeof value === 'string' ? parseISO(value) : value;
-
-    if(value) {
-      const date = new Date(value);
-      const timezoneOffset = date.getTimezoneOffset();
-      const offsetHours = Math.floor(Math.abs(timezoneOffset) / 60);
-      const offsetMinutes = Math.abs(timezoneOffset) % 60;
-      const offsetSign = timezoneOffset < 0 ? "+" : "-";
-      const offsetFormatted = `${offsetSign}${String(offsetHours).padStart(2, "0")}:${String(offsetMinutes).padStart(2, "0")}`;
-      const adjustedTimestampWithoutZ = new Date(date.getTime() - timezoneOffset * 60 * 1000).toISOString().slice(0, -1);
-      const adjustedTimestampWithOffset = `${adjustedTimestampWithoutZ}${offsetFormatted}`;
-      // console.log('adjustedTimestampWithOffset',adjustedTimestampWithOffset);
-      dateValue = adjustedTimestampWithOffset;
-    }
-  
-  // console.log('dateValue', dateValue);
+    
   /* Note: the wrapper div is there because MuiPicker creates two child elements,
    * which messes up display if this component is a flex child. */
   return (
