+++ src/pages/sighting/EditSightingMetadata.jsx
@@ -37,8 +37,6 @@ export default function EditSightingMetadata({
     clearError: clearSightingError,
   } = usePatchSighting();
 
-  // console.log('sighting error',sightingError);
-
   const {
     mutate: updateAgsProperties,
     loading: agsLoading,
